Move dhis2Server.ts to src/utils
